+++ src/kptr.jl
@@ -35,7 +35,8 @@ end
 function freeKnetPtr(p::KnetPtr)
     if p.ptr == C_NULL || isdefined(p,:parent); return; end
     mem = KnetMems[p.dev+1]
-    mem.avail += p.len
+    mem.bfree += p.len
+    mem.kfree += 1
     push!(mem.pools[p.len].free, p.ptr)
     p.ptr = C_NULL # to avoid double free by gcnode then gc.
 end
@@ -56,13 +57,19 @@ mutable struct KnetMem
     pools::Dict{Int,KnetPool}   # pointers of a given size
     limit::Int                  # current memory limit
     bytes::Int                  # total bytes allocated (inuse + avail)
-    avail::Int                  # total bytes freed and available
-    KnetMem()=new(Dict{Int,KnetPool}(),KNETMEMINIT,0,0)
+    bfree::Int                  # total bytes freed and available
+    kptrs::Int                  # number of arrays allocated (inuse + avail)
+    kfree::Int                  # number of arrays freed and available
+    gc::Int                     # number of times GC.gc called
+    knetgc::Int                 # number of times Knet.gc called
+    gctime::UInt
 end
+KnetMem()=KnetMem(Dict{Int,KnetPool}(),KNETMEMINIT,0,0,0,0,0,0,0)
 
 # KnetMems[dev+1] holds memory information for device dev.
 KnetMems = nothing
 initKnetMems() = (global KnetMems = [ KnetMem() for i in 1:gpuCount() ])
+knetmem(dev=gpu()) = (if KnetMems == nothing; initKnetMems(); end; KnetMems[dev+1])
 
 # Blocksize determines the actual allocation size given the array size in bytes, and can be
 # larger than what the array needs for increased reuse.
@@ -77,71 +84,92 @@ blocksize(n::Int,b=cbrt(2))=floor(Int,b^ceil(log(b,n)))
 # The following used for debugging and record every request
 arraysizes = Int[]; allocs = Int[]; blocksizes = Int[]
 
-function inclimit!(m::KnetMem, minlimit=m.limit*6÷5)
-    maxlimit = gpumem()[1] - 500_000_000 # m.bytes + gpufree()
-    minlimit = min(maxlimit, minlimit)
-    if m.limit < minlimit
-        m.limit = minlimit
-        @debug "limit=$(m.limit) bytes=$(m.bytes) avail=$(m.avail) pools=$(length(m.pools)) blocks=$(sum(p->p.nptr,values(m.pools)))"
-    end
-end
+# This the main KnetPtr constructor.  It tries to avoid actual allocation which is slow.
+# Reusing a preallocated and garbage collected pointer is very fast.
+# Allocating a new pointer is about 0.5ms.
+# GC.gc() is about 100ms. 
+# Knet.gc() is about 250ms.
 
-# This the main KnetPtr constructor.  It tries to avoid actual
-# allocation which is slow.  Reusing a pointer is very fast.
-# Allocating a new one is about 10 μs.  gc() is about 75 ms.
+gc_interval() = 2*10^8  # gc interval in ns, optimized on seq2seq model, balancing costs of alloc, GC.gc, Knet.gc
+putc(c)=nothing         # putc(c)=print(c) to observe GC.gc, Knet.gc and inclimit
 
 function KnetPtr(arraybytes::Int)
-    dev = gpu(); if dev < 0; error("KnetPtr: bad device id $dev."); end
-    if KnetMems==nothing; initKnetMems(); end
-    mem = KnetMems[dev+1]
+    dev = gpu(); @assert dev >= 0 "KnetPtr: bad device id $dev."
+    mem = knetmem(dev)
     blockbytes = blocksize(arraybytes)
     @dbg (push!(arraysizes,arraybytes); push!(blocksizes,blockbytes))
     pool = get!(KnetPool,mem.pools,blockbytes)
-    if !isempty(pool.free)      # 1. best case we have one available in pool
-        @dbg push!(allocs, 1)
-        mem.avail -= blockbytes
-        return KnetPtr(pop!(pool.free),blockbytes,dev)
+
+    ptr = reuse(mem, pool, blockbytes, 1) # 1. best case we have one available in pool
+    ptr != nothing && return KnetPtr(ptr,blockbytes,dev)
+    if pool.nptr > 0 && time_ns() - mem.gctime > gc_interval() # 2. try gc (~100 ms) if we had allocated this size before and enough time passed
+        ptr = reuse(mem, pool, blockbytes, 2, trygc=true); putc('-')
+        ptr != nothing && return KnetPtr(ptr,blockbytes,dev)
     end
-    if mem.bytes + blockbytes <= mem.limit # 2. allocate if we are within limit
-        ptr = knetMalloc(blockbytes)
-        if ptr != nothing
-            @dbg push!(allocs, 2)
-            mem.bytes += blockbytes
-            pool.nptr += 1
-            return KnetPtr(ptr,blockbytes,dev)
-        end
+    if mem.bytes + blockbytes <= mem.limit # 3. allocate if within limit
+        ptr = alloc(mem, pool, blockbytes, 3)
+        ptr != nothing && return KnetPtr(ptr,blockbytes,dev)
     end
-    if pool.nptr > 0            # 3. try gc if we had allocated this size before
-        GC.gc();  @dbg print('-')
-        # inclimit!(mem, (mem.bytes-mem.avail)*3) # *4=>18s *3=>16s *5÷2=>16s *2=>22s *6÷5=>30s
-        inclimit!(mem, mem.limit*6÷5)
-        if !isempty(pool.free)
-            @dbg push!(allocs, 3)
-            mem.avail -= blockbytes
-            return KnetPtr(pop!(pool.free),blockbytes,dev)
-        end
+    if maybe_inclimit!(mem, max(mem.bytes + blockbytes*2, mem.limit*6÷5)) # 4. try to increase limit
+        ptr = alloc(mem, pool, blockbytes, 4); putc('^')
+        ptr != nothing && return KnetPtr(ptr,blockbytes,dev)
     end
-    # At this point we have to either free old ptrs or increase limit
-    # Let's just try limit increase for now
-    inclimit!(mem, (mem.bytes + blockbytes*2))
-    ptr = knetMalloc(blockbytes)
-    if ptr == nothing
-        Knet.gc(); @dbg print('+')  # last ditch effort
-        ptr = knetMalloc(blockbytes)
-    end
-    if ptr != nothing
-        @dbg push!(allocs, 4)
-        mem.bytes += blockbytes
-        pool.nptr += 1
-        return KnetPtr(ptr,blockbytes,dev)
+    ### This does not prevent Knet.gc for too long:
+    # if pool.nptr > 0 && time_ns() - mem.gctime > gc_interval2()  # 5. One last gc before Knet.gc bomb
+    #     ptr = reuse(mem, pool, blockbytes, 5, trygc=true); putc('=')
+    #     ptr != nothing && return KnetPtr(ptr,blockbytes,dev)
+    # end
+    Knet.gc(); putc('+')  # 6. last ditch effort: ~250 ms + future cost of lost pools
+    if mem.bytes + blockbytes <= mem.limit
+        ptr = alloc(mem, pool, blockbytes, 6)
+        ptr != nothing && return KnetPtr(ptr,blockbytes,dev)
     end
     error("Out of gpu memory")
+end
+
+function alloc(mem, pool, blockbytes, dbg)
+    ptr = knetMalloc(blockbytes)       # ~584μs
+    if ptr != nothing
+        @dbg push!(allocs, dbg)
+        mem.bytes += blockbytes
+        mem.kptrs += 1
+        pool.nptr += 1
+        return ptr
+    else
+        return nothing
+    end
+end
+
+function reuse(mem, pool, blockbytes, dbg; trygc=false)
+    if trygc
+        @timeit to "gc" GC.gc()
+        mem.gctime = time_ns(); mem.gc += 1
+    end
+    if !isempty(pool.free)
+        @dbg push!(allocs, dbg)
+        mem.bfree -= blockbytes
+        mem.kfree -= 1
+        return pop!(pool.free)
+    else
+        return nothing
+    end
+end
+
+function maybe_inclimit!(m::KnetMem, minlimit=m.limit*6÷5)
+    maxlimit = m.bytes + gpufree() - 500_000_000 # gpumem()[1] - 500_000_000 # m.bytes + gpufree()
+    if m.limit < minlimit <= maxlimit
+        m.limit = minlimit
+        @debug kmeminfo()
+        return true
+    else
+        return false
+    end
 end
 
 KnetPtr(n::Integer)=KnetPtr(Int(n))
 
 # This does the actual allocation, returns `nothing` in case of error
-function knetMalloc(nbytes::Int)
+function knetMalloc(nbytes::Int) # 584μs
     ptr = Cptr[0]
     ret = @cudart1(cudaMalloc,(Ptr{Cptr},Csize_t),ptr,nbytes)
     ret == 0 ? ptr[1] : nothing
@@ -157,14 +185,17 @@ you run out of GPU memory.
 """
 function gc(dev=gpu())
     if KnetMems == nothing; return; end
-    mem = KnetMems[dev+1]
+    mem = knetmem(dev)
+    mem.knetgc += 1
     GC.gc(); GC.enable(false)
     for (n,v) in mem.pools
         for p in v.free
             @cudart(cudaFree,(Cptr,),p)
         end
         v.nptr -= length(v.free)
-        mem.avail -= n * length(v.free)
+        mem.kptrs -= length(v.free)
+        mem.kfree -= length(v.free)
+        mem.bfree -= n * length(v.free)
         mem.bytes -= n * length(v.free)
         empty!(v.free)
     end
@@ -177,12 +208,16 @@ function knetgc()
 end
 
 # Some utilities
-meminfo(i=gpu())=(KnetMems==nothing ? [] : [(k,v.nptr,length(v.free)) for (k,v) in KnetMems[i+1].pools])
+
+meminfo(i=gpu())=[(k,v.nptr,length(v.free)) for (k,v) in knetmem(i).pools]
+
+using Printf
+kmeminfo(i=gpu())=(m=knetmem(i); @sprintf("knetgc=%g gc=%g pools=%g kptrs=%g kfree=%g limit=%g bytes=%g bfree=%g", m.knetgc, m.gc, length(m.pools), m.kptrs, m.kfree, m.limit, m.bytes, m.bfree))
 
 function gpuinfo(msg="",dev=gpu();n=10)
     msg != "" && print("$msg ")
     if nvmlfound # Libdl.find_library(["libnvidia-ml"],[]) != ""
-        g = nvmlDeviceGetMemoryInfo(dev)
+        g = nvmlDeviceGetMemoryInfo(nvmlid(dev))
         println((:dev,dev,:total,g[1],:free,g[2],:used,g[3]))
     else
         dev0 = gpu(); gpu(dev)
@@ -190,16 +225,16 @@ function gpuinfo(msg="",dev=gpu();n=10)
         gpu(dev0)
         println((:dev,dev,:total,total,:free,free,:used,total-free))
     end
-    bytes = avail = ptrs = k = 0
+    bytes = bfree = ptrs = k = 0
     for (s,u,f) in sort(meminfo(dev), by=(x->x[1]*x[2]), rev=true)
-        bytes += s*u; avail += s*f; ptrs += u; k += 1
-        if k <= n; println((:bytes,s*u,:size,s,:alloc,u,:avail,f)); end
+        bytes += s*u; bfree += s*f; ptrs += u; k += 1
+        if k <= n; println((:bytes,s*u,:size,s,:alloc,u,:bfree,f)); end
     end
     if n < k; println('⋮'); end
-    println((:totbytes,bytes,:avail,avail,:ptrs,ptrs))
+    println((:totbytes,bytes,:bfree,bfree,:ptrs,ptrs))
     if KnetMems != nothing
         mem = KnetMems[dev+1]
-        println((:membytes,mem.bytes,:avail,mem.avail,:limit,mem.limit,:pools,length(mem.pools)))
+        println((:membytes,mem.bytes,:bfree,mem.bfree,:limit,mem.limit,:pools,length(mem.pools)))
     end
 end
 
@@ -222,3 +257,4 @@ KnetPtr: ktotal: $(x[1]) kavail: $(x[2]) ktotal-kavail: $(x[1]-x[2])
 nused-ktotal: $(m[3]-x[1])
 """)
 end
+