+++ test/bmm.jl
@@ -3,10 +3,10 @@ using Statistics, LinearAlgebra
 
 sizes = [((2,4,3),(4,1,3)),((2,4,5),(4,8,5)),((2,8,4),(8,2,4))]
 @testset "bmm" begin
-    for t in (Float32,Float64)
+    for t in (Float32, Float64)
         for s in sizes
-            a = 0.1*randn(t, s[1]...)
-            b = 0.1*randn(t, s[2]...)
+            a = t(0.1)*randn(t, s[1]...)
+            b = t(0.1)*randn(t, s[2]...)
             bmmul(w)=bmm(w[1],w[2])
             @test gradcheck(bmmul, (a,b))
             if gpu() >= 0